feat: consolidate banners into single component
