+++ apps/cowswap-frontend/src/modules/tokensList/pure/TokensVirtualList/index.tsx
@@ -5,7 +5,7 @@ import { TokenWithLogo } from '@cowprotocol/common-const'
 import { useVirtualizer } from '@tanstack/react-virtual'
 import ms from 'ms.macro'
 
-import { CoWAmmTokenSelectorBanner } from 'common/pure/CoWAMMBanner/tokenSelectorBanner'
+import { CoWAmmBanner } from 'common/pure/CoWAMMBanner'
 
 import * as styledEl from './styled'
 
@@ -67,7 +67,7 @@ export function TokensVirtualList(props: TokensVirtualListProps) {
 
   return (
     <CommonListContainer id="tokens-list" ref={parentRef} onScroll={onScroll}>
-      <CoWAmmTokenSelectorBanner />
+      <CoWAmmBanner location="tokenSelector" />
 
       <styledEl.TokensInner ref={wrapperRef} style={{ height: virtualizer.getTotalSize() }}>
         <styledEl.TokensScroller style={{ transform: `translateY(${items[0]?.start ?? 0}px)` }}>
